requirements.txt
pytest==8.3.3
selenium==4.27.1
allure-python-commons~=2.13.5